requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==4.1.2
django-filter==22.1
django-taggit==1.3.0
djangorestframework==3.14.0
Markdown==3.4.1
Pillow==9.2.0
pycodestyle==2.9.1
pytz==2022.4
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.4
